diamond is now visually dragged

+++ src/shared.rs
@@ -620,7 +620,7 @@ pub struct UiAnim {
     pub exported_frame: String,
 
     pub timeline_offset: Vec2,
-    pub dragged_keyframe: usize,
+    pub dragged_keyframe: i32,
     pub icon_images: Vec<egui::TextureHandle>,
     pub loops: i32,
